Add RWStorage buffers to reflection and ability to bind in command buffer
Metadata can cope with structured buffers that have a struct and those that have a basic type
Basic types less than the SIMD width won't work properly as we're mapping types to hlsl++ which always uses __m128 as the underlying type. This is not such a bad thing for constant buffers, where being aligned to multiples of 4 floats is a good idea anyway.
For structured buffers we have two options: 1) actual data types that correspond to shader sizes, (perhaps meta::float3 or something similar). These types could be easily transformed from hlsl++ types for seamless integration. 2) To simply return the underlying type, like float or int, and the calling code that populates the buffer takes care of the number of elements
For structured buffers it's also not a bad idea to have data in multiples of 4 float because loads will happen at that granularity