Replace gainput with SDL and rework CrInputManager to accommodate
The CrInput model is such that handlers feed into the input manager, and the input manager propagates it to the rest of the codebase. The advantages of this model is that handlers can be created in any shape or form, as long as they feed in the data that the input manager expects. For example, a replay handler could record input commands to a text file, and simply feed them back in. It also plays nicely with event-based input vs poll-based input, as each handler can do whatever approach fits best, which is important as consoles and desktop often can work differently
Imgui now has full support for keyboard and mouse events
Controller navigation has been reworked and mouse navigation has very primitive support now