Add solution to GPU resource deletion
There is now a deletion queue object, that the render device currently owns. During the frame we collect resources for deletion and they have a GPU fence associated with them. At the end of the frame we move the current list to an active list and schedule the fence for signaling on the GPU. This model assumes that any resource deleted during the frame is potentially being used in command buffers up to the point where we schedule the fence, which may or may not be true. However, this model seems better than assigning a number of frames for deletion to happen as it makes no assumptions about the behavior of the frame or the time it takes to do something.
Only some resources currently are assigned to this queue, we will improve it to make it work for all of them eventually