Add CrPathString
We are replacing the CrPath alias of std::filesystem::path as the default way of doing paths due to the overhead and dynamic allocation of creating strings out of them
The new CrPath is a fixed size string which can be more memory but does no allocations
The interface is very similar to std::filesystem::path although the behavior can be a bit different (e.g. regarding the / operator)
The commit right after will rename it to CrPath
If the size of the path becomes a problem, we could have a short path and a long path. One would be useful for paths relative to the root of the data, the other useful for absolute paths when going to fetch data from the actual filesystem. The current path would be the long one, the other an optimization