Add initial stage of the render world
Model Instances are represented by a model instance id. All the data lives inside the render world
Data is contiguous in the render world so that visibility can be computed very efficiently later
This means there is an indirection to go from model instance to its data, but that is an operation that isn't frequent beyond the initial setup
Added a first visibility function. The next step is to process visibility and create rendering packets from it. The rendering never iterates through model instances, instead it works on more granular primitives