Added render graph
The render graph is a flexible way of tracking resource usage throughout a frame, among other things
The render graph invokes the platform-specific BeginRenderPass and EndRenderPass, and each platform is responsible for translating the commands appropriately
As APIs get implemented, the abstraction probably needs adjusting but it should be fairly flexible
There are things missing such as the scope where the resource is going to be used (in the case of a texture or RWtexture)
Some platforms may not need transitions but instead flush caches, others can be more fine-grained
Render passes now begin a debug event as well, as we want to make sure the event is fired before the render pass begins (and we can capture all commands within)
Clean up all the texture barrier API which we aren't going to use anymore
One API we will need is a memory dependency for dispatches inside a compute pass. Resources don't necessarily transition, they just need to wait for each other. The same effect can be achieved through multiple compute passes, but sometimes specifying a manual wait keeps code simpler