Update D3D12 backend which is now on par with the Vulkan backend
Added root signature management (still needs work as Xbox platforms need the root signature to be specified at compile time)
Added descriptor heap management and upload for graphics shaders (compute still missing). This lays down the work needed for resource tables, also the model for it is to create resource views for CBV, SRV, UAV and samplers on the fly through the descriptor instead of creating the descriptor handle "offline" like in D3D11. This means we don't get many random copies to GPU, instead we have these fast API calls
Bind vertex, index buffers, viewports and all sorts of dynamic state in the graphics flush
Create and manage views and texture upload in the texture class. This upload is still very much a work in progress as it needs a better API (so does Vulkan)