Add Editor Edge Selection shader code and render pass
Only triggered when left mouse button is clicked
We need to be able to download the instance id (editor only) buffer to determine which instance to highlight
Crate constant buffers before the pass. That way we only lock and bind once
Fix the GPU Timing Query Tracker. It wasn't clearing the hashmap properly
Fix Copy Pass. It was trying to load the contents of an undefined texture
Also rework the material define generator. We want to be able to arbitrarily define enums in shader code