Introduce model instance selection
We can now select model instances, multi-select with shift, and clear via escape. This needed a rework on the RenderWorld side
Added initial data for buffers in the same way we do textures
Added render device support to upload data to buffers, via BeginBufferUpload and EndBufferUpload. This API ensure we can abstract away differences between PC and console, as internally we create staging buffers or not depending on platform
Added render device support to download buffer data, and a response lambda via a callback, which allows us to read the contents of the entity selection id
Change the name of the buffer resource to just d3d12Resource
Added debugging names for buffers
Added D3D12Cast and VulkanCast to try to reduce casting clutter