Rework GPU Buffers
We now have the concept of a GPU Buffer View. This view is not unlike the API views, but these are short lived, don't create API objects and are mainly for binding and data uploading. We split it off of the GPU Buffer class because many concepts were getting confused. Crucially the ownership of the buffer was getting awkward to manage
GPU Buffers are now a wrapper around a hardware buffer with some extra data. They cover the entire buffer and cannot point to part of another buffer. We unfortunately cannot derive from the hardware GPU buffer so we own one instead. The GPU buffer is a bit like a special incarnation of the hardware GPU buffer, for when we want vertex descriptors, index buffer formats, data formats, etc
We've had to change the render graph buffer declaration to accommodate the new model, the previous one wasn't very flexible anyway
We've also had to change the render graph descriptor to take hardware buffers, size and offset
Buffer views now have a GetData() function instead of Lock(), Unlock(), to further cement the idea that they don't own any data or map a buffer, that has to have happened elsewhere