Move all rendering resources to be intrusive_ptr, and remove all instances of eastl::shared_ptr
intrusive_ptr has some advantages with regards to memory allocation, inlining and memory that make it suitable. We may reintroduce a more lightweight shared_ptr down the line
Fix the deletion of fences, command buffers and other resources via the deletion list
Rework the intrusive_ptr to call a templated deletion callback, so we can cast it to the right type
Add m_renderDevice to the GPUDeletable class. Delete all other render devices from other resources. Eventually we may just have a static, it's hard to say at the moment how to create two separate pipelines so that Vulkan and D3D12 can run in parallel