Replace eastl::string with crstl::string
crstl::string and fixed_string are usable and have been polished as much as possible through normal usage in the engine
This of course doesn't mean that they are perfect, but at least is at a point where adding extra functionality and refining performance is not hard to do
Remove typedefs to eastl shared_ptr, unique_ptr, etc
Replace to_string with the string constructor version. The reason for this is that to_string assumes we want to construct a string, but we could want a fixed_string
Renamed force_size to force_length to be consistent with the rest of the naming for string. This is used for example when reading into a string from a file. A better API would be to pass in a capacity and a default length