Rework how we print out constant buffer metadata
We no longer have 2 structures, e.g. LightingData and Lighting, instead we have a single one named after the constant buffer. We inline the size of the struct and static assert if the size calculated by C++ and the one provided by reflection are the same, this way we detect padding issues